revising sizing and creating saved airport
