+++ server/public/media.css
@@ -130,4 +130,11 @@ form.airportForm{
   width:30%;
 }
 
+h6{
+  font-size:1.2rem;
+}
+h5{
+  font-size:.9rem;
+}
+
 }
